api/requirements.txt
flask
flask-cors
pandas
scikit-learn
spacy
pdfplumber
python-docx
